Moved responssability to eventhandler
