+++ src/main/java/com/ticketpurchasingsystem/project/application/IActiveOrderService.java
@@ -15,6 +15,6 @@ public interface  IActiveOrderService {
     public void addStandingAreaToActiveOrder(SessionToken sessionToken, String orderId, String areaId, int quantity);
     public void updateActiveOrder(SessionToken sessionToken, ActiveOrderDTO order);
     public boolean saveOrder(ActiveOrderItem order);
-    public List<BarcodeDTO> completeOrder(IPaymentGateway paymentGateway, SessionToken sessionToken, double amount, String orderId);
+    public List<BarcodeDTO> completeOrder(IPaymentGateway paymentGateway, SessionToken sessionToken, PaymentDetails paymentDetails, String orderId);
     public ActiveOrderDTO getActiveOrderByUserId(SessionToken sessionToken, String userId) throws Exception;
 }
